log + cmd bug fixes
Small fixes spotted when compiling for MAC & Android.

* Print-out format fixes. Casts have been introduced at some places to clarify the matching type and satisfy tested compilers.
* Update Android makefile with new cmd parsing component.
* Wrap getpid() with arch defines, enabling syscall only for LINUX arch.
* Clean-up unused headers in cmdline.c
* Split LOG_HELP print messages since clang was complaining:
cmdline.c:88:2: error: embedding a directive within macro arguments has undefined behavior [-Werror,-Wembedded-directive]

+++ linux/perf.c
@@ -267,7 +267,7 @@ static bool arch_perfOpen(pid_t pid, dynFileMethod_t method, int *perfFd)
                      MAP_PRIVATE | MAP_ANONYMOUS | MAP_NORESERVE, -1, 0);
     if (perfBloom == MAP_FAILED) {
         perfBloom = NULL;
-        PLOG_E("mmap(size=%zu) failed", _HF_PERF_BLOOM_SZ);
+        PLOG_E("mmap(size=%zu) failed", (size_t) _HF_PERF_BLOOM_SZ);
     }
 
     perfMmapBuf =
